requirements.txt
praw>=7.7.1
Pillow>=10.0.0
PyYAML>=6.0.1
requests>=2.31.0
